# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Anima-Link API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  backend: string;
  llm_configured: boolean;
  sqlite_connected: boolean;
  chroma_connected: boolean;
  mode: string;
}

export interface AnalyzeMessageRequest {
  message: string;
  user_id?: string;
  session_id?: string;
}

export interface MemoryItem {
  id: string;
  title: string;
  summary: string;
  trigger: string;
  emotion: string;
  distortion?: string;
  reframe: string;
  character: string;
  created_at: string;
}

export interface MessageAnalysis {
  analysis_id: string;
  message_id: string;
  emotion: string;
  msi_score: number;
  msi_label: string;
  absolutist_words: string[];
  distortion: string;
  trigger: string;
  intent: string;
  recommended_path: string[];
  memory_query: string[];
  suggested_game: string;
  safe_response: string;
  memory_results: MemoryItem[];
}

export interface AddMemoryRequest {
  title: string;
  content: string;
  emotion: string;
  trigger: string;
  distortion?: string;
  reframe?: string;
  character?: string;
  memory_type?: string;
}

export interface AddMemoryResponse {
  id: string;
  success: boolean;
}

export interface SearchMemoryRequest {
  query: string;
  category?: string;
  limit?: number;
}

export interface MemorySearchResponse {
  query: string;
  results: MemoryItem[];
  source: string;
}

export interface AllMemoriesResponse {
  memories: MemoryItem[];
  total: number;
}

export interface MemoryGraphNode {
  id: string;
  label: string;
  type: string;
  emotion: string;
  date: string;
}

export interface MemoryGraphEdge {
  from: string;
  to: string;
  relation: string;
}

export interface MemoryGraph {
  nodes: MemoryGraphNode[];
  edges: MemoryGraphEdge[];
}

export type CharacterRouteRequestAnalysis = { [key: string]: unknown };

export type CharacterRouteRequestMemoryResultsItem = { [key: string]: unknown };

export interface CharacterRouteRequest {
  analysis: CharacterRouteRequestAnalysis;
  memory_results?: CharacterRouteRequestMemoryResultsItem[];
  user_preference?: string;
}

export interface CharacterResponse {
  name: string;
  role: string;
  message: string;
  reason: string;
}

export interface CharacterRouteResponse {
  recommended_path: string[];
  primary_character: string;
  reason: string;
  character_responses: CharacterResponse[];
}

export type ThoughtMonsterStartRequestMemoryResultsItem = { [key: string]: unknown };

export interface ThoughtMonsterStartRequest {
  message: string;
  analysis_id?: string;
  emotion?: string;
  distortion?: string;
  absolutist_words?: string[];
  memory_results?: ThoughtMonsterStartRequestMemoryResultsItem[];
}

export type ThoughtMonsterGameMemoryProof = { [key: string]: unknown } | null;

export interface ThoughtMonsterGame {
  game_id: string;
  monster_name: string;
  trigger_words: string[];
  distortion: string;
  weakness: string;
  guide: string;
  cbt_question_1: string;
  cbt_question_2: string;
  cbt_question_3: string;
  memory_proof?: ThoughtMonsterGameMemoryProof;
  recommended_next_action: string;
  safe_reframe: string;
}

export interface ThoughtMonsterCompleteRequest {
  game_id: string;
  original_thought: string;
  final_reframe: string;
  distortion?: string;
  character?: string;
}

export interface ThoughtMonsterResult {
  success: boolean;
  reward: string;
  memory_id?: string;
  graph_updated: boolean;
}

export interface CameraAnalyzeRequest {
  image_base64: string;
  mode?: string;
}

export interface WorkspaceObject {
  label: string;
  game_label: string;
  category: string;
  confidence: string;
}

export interface CameraAnalysisResult {
  objects: WorkspaceObject[];
  workspace_state: string;
  suggested_mission: string;
  focus_score: number;
  source: string;
}

export interface FocusBossCreateRequest {
  goal: string;
  linked_thought?: string;
  camera_objects?: string[];
  estimated_minutes?: number;
}

export interface FocusBossStep {
  step: number;
  action: string;
  duration_minutes?: number;
}

export interface FocusBossPlan {
  boss_id: string;
  boss_name: string;
  main_task: string;
  tiny_step: string;
  estimated_time_minutes: number;
  battle_plan: FocusBossStep[];
  character: string;
  arlo_message: string;
}

export interface TinyWinRequest {
  proof_type: string;
  completed_action: string;
  linked_thought?: string;
  counter_evidence?: string;
  character?: string;
}

export interface TinyWinResponse {
  id: string;
  artifact: string;
  reward_message: string;
  graph_updated: boolean;
  memory_id?: string;
}

export interface WeeklyReportSession {
  date: string;
  emotion: string;
  distortion: string;
  character: string;
  intervention: string;
}

export interface WeeklyReport {
  period: string;
  total_sessions: number;
  dominant_emotion: string;
  most_common_distortion: string;
  most_effective_character: string;
  tiny_wins_count: number;
  thought_battles_won: number;
  msi_average: number;
  progress_summary: string;
  sessions: WeeklyReportSession[];
  recommendations: string[];
}

export interface SafetyCheckRequest {
  message: string;
  user_id?: string;
}

export interface SafetyCheckResponse {
  safe: boolean;
  risk_level: string;
  route: string;
  message: string;
  resources: string[];
}

export interface VoiceTranscribeRequest {
  audio_base64?: string;
  typed_fallback?: string;
}

export interface VoiceTranscribeResponse {
  transcript: string;
  source: string;
  confidence: string;
}

export interface VoiceRouteRequest {
  transcript: string;
}

export type VoiceRouteResponseParams = { [key: string]: unknown };

export interface VoiceRouteResponse {
  route: string;
  action: string;
  params?: VoiceRouteResponseParams;
}