# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Anima-Link API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memory/add:
    post:
      operationId: addMemory
      tags: [memory]
      summary: Add a memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemoryRequest"
      responses:
        "200":
          description: Memory added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddMemoryResponse"
  /memory/search:
    post:
      operationId: searchMemory
      tags: [memory]
      summary: Search memories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchMemoryRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemorySearchResponse"
...
  /games/thought-monster/start:
    post:
      operationId: startThoughtMonster
      tags: [games]
      summary: Start a Thought Monster Battle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThoughtMonsterStartRequest"
      responses:
        "200":
          description: Game started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThoughtMonsterGame"
  /games/thought-monster/complete:
    post:
      operationId: completeThoughtMonster
      tags: [games]
      summary: Complete a Thought Monster Battle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThoughtMonsterCompleteRequest"
      responses:
        "200":
          description: Game completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThoughtMonsterResult"
...
  /games/focus-boss/create:
    post:
      operationId: createFocusBoss
      tags: [games]
      summary: Create a Focus Boss Fight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FocusBossCreateRequest"
      responses:
        "200":
          description: Focus boss plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusBossPlan"
  /tiny-win/store:
    post:
      operationId: storeTinyWin
      tags: [tiny-win]
      summary: Store a tiny win proof
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TinyWinRequest"
      responses:
        "200":
          description: Tiny win stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TinyWinResponse"
...
  /voice/transcribe:
    post:
      operationId: transcribeVoice
      tags: [voice]
      summary: Transcribe audio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceTranscribeRequest"
      responses:
        "200":
          description: Transcription result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceTranscribeResponse"
  /voice/route-spell:
    post:
      operationId: routeVoiceSpell
      tags: [voice]
      summary: Route voice command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceRouteRequest"
      responses:
        "200":
          description: Voice routing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceRouteResponse"
components:
  schemas:
...
    MemoryItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        trigger:
          type: string
        emotion:
          type: string
        distortion:
          type: string
        reframe:
          type: string
        character:
          type: string
        created_at:
          type: string
      required:
        - id
        - title
        - summary
        - trigger
        - emotion
        - reframe
        - character
        - created_at
    AddMemoryRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        emotion:
          type: string
        trigger:
          type: string
        distortion:
          type: string
        reframe:
          type: string
        character:
          type: string
        memory_type:
          type: string
      required:
        - title
        - content
        - emotion
        - trigger
    AddMemoryResponse:
      type: object
      properties:
        id:
          type: string
        success:
          type: boolean
      required:
        - id
        - success
    SearchMemoryRequest:
      type: object
      properties:
        query:
          type: string
        category:
          type: string
        limit:
          type: integer
      required:
        - query
    MemorySearchResponse:
      type: object
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/MemoryItem"
        source:
          type: string
      required:
        - query
        - results
        - source
...
    ThoughtMonsterStartRequest:
      type: object
      properties:
        message:
          type: string
        analysis_id:
          type: string
        emotion:
          type: string
        distortion:
          type: string
        absolutist_words:
          type: array
          items:
            type: string
        memory_results:
          type: array
          items:
            type: object
      required:
        - message
    ThoughtMonsterGame:
      type: object
      properties:
        game_id:
          type: string
        monster_name:
          type: string
        trigger_words:
          type: array
          items:
            type: string
        distortion:
          type: string
        weakness:
          type: string
        guide:
          type: string
        cbt_question_1:
          type: string
        cbt_question_2:
          type: string
        cbt_question_3:
          type: string
        memory_proof:
          type: object
          nullable: true
        recommended_next_action:
          type: string
        safe_reframe:
          type: string
      required:
        - game_id
        - monster_name
        - trigger_words
        - distortion
        - weakness
        - guide
        - cbt_question_1
        - cbt_question_2
        - cbt_question_3
        - recommended_next_action
        - safe_reframe
    ThoughtMonsterCompleteRequest:
      type: object
      properties:
        game_id:
          type: string
        original_thought:
          type: string
        final_reframe:
          type: string
        distortion:
          type: string
        character:
          type: string
      required:
        - game_id
        - original_thought
        - final_reframe
    ThoughtMonsterResult:
      type: object
      properties:
        success:
          type: boolean
        reward:
          type: string
        memory_id:
          type: string
        graph_updated:
          type: boolean
      required:
        - success
        - reward
        - graph_updated
...
    FocusBossCreateRequest:
      type: object
      properties:
        goal:
          type: string
        linked_thought:
          type: string
        camera_objects:
          type: array
          items:
            type: string
        estimated_minutes:
          type: integer
      required:
        - goal
    FocusBossStep:
      type: object
      properties:
        step:
          type: integer
        action:
          type: string
        duration_minutes:
          type: integer
      required:
        - step
        - action
    FocusBossPlan:
      type: object
      properties:
        boss_id:
          type: string
        boss_name:
          type: string
        main_task:
          type: string
        tiny_step:
          type: string
        estimated_time_minutes:
          type: integer
        battle_plan:
          type: array
          items:
            $ref: "#/components/schemas/FocusBossStep"
        character:
          type: string
        arlo_message:
          type: string
      required:
        - boss_id
        - boss_name
        - main_task
        - tiny_step
        - estimated_time_minutes
        - battle_plan
        - character
        - arlo_message
    TinyWinRequest:
      type: object
      properties:
        proof_type:
          type: string
        completed_action:
          type: string
        linked_thought:
          type: string
        counter_evidence:
          type: string
        character:
          type: string
      required:
        - proof_type
        - completed_action
    TinyWinResponse:
      type: object
      properties:
        id:
          type: string
        artifact:
          type: string
        reward_message:
          type: string
        graph_updated:
          type: boolean
        memory_id:
          type: string
      required:
        - id
        - artifact
        - reward_message
        - graph_updated
...
    VoiceTranscribeRequest:
      type: object
      properties:
        audio_base64:
          type: string
        typed_fallback:
          type: string
      required: []
    VoiceTranscribeResponse:
      type: object
      properties:
        transcript:
          type: string
        source:
          type: string
        confidence:
          type: string
      required:
        - transcript
        - source
        - confidence
    VoiceRouteRequest:
      type: object
      properties:
        transcript:
          type: string
      required:
        - transcript
    VoiceRouteResponse:
      type: object
      properties:
        route:
          type: string
        action:
          type: string
        params:
          type: object
      required:
        - route
        - action