# lib/api-zod/src/generated/types/voiceRouteResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Anima-Link API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VoiceRouteResponseParams } from './voiceRouteResponseParams';

export interface VoiceRouteResponse {
  route: string;
  action: string;
  params?: VoiceRouteResponseParams;
}